refactor(api-go): split s3 multipart handlers
